refactor(init): split original project to separate concerns
This project was the original data collector for project
'Baronerd' which was renamed to project 'Air' within this work.
Along with the rename, the new project was split up into
multiple git repositories in order to separate development and
running environment concerns. This project is only concerned
with the data collection portion of 'Air', leaving plotting of
the data and other functionality to the other repos within
this project.