feat!(api data): change DarkSky for ClimaCell data
Change data source for weather data from DarkSky API to
ClimaCell API due to sunset of DarkSky API access, along
with the fact that ClimaCell also supplies air quality and
pollen data

BREAKING CHANGE: air_collect prior to this change with depend
on the DarkSky api data source for weather data.  With this change
ClimaCell api is used instead and no the DarkSky api is not used
for any data.